% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_penguin_doc.R
\docType{data}
\name{pal_penguin}
\alias{pal_penguin}
\title{Sediment trap in nearshore waters collected during Palmer LTER station season at Palmer Station Antarctica, 1992 - 1995.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 335 rows and 12 columns.
}
\source{
{Palmer Station Antarctica LTER and D. Karl. 2017. Sediment trap in nearshore waters collected during Palmer LTER station season at Palmer Station Antarctica, 1992 - 1995. ver 3. Environmental Data Initiative. https://doi.org/10.6073/pasta/816f47788862f4dfc4437333e081bfb8 (Accessed 2021-02-22).}
\url{https://portal.edirepository.org/nis/mapbrowse?packageid=knb-lter-pal.46.3}
}
\usage{
pal_penguin
}
\description{
Particulate organic matter is exported from the upper ocean euphotic zone in the form of large sinking particles and as dissolved material.
}
\keyword{datasets}
